backend/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jsr-hydra"
version = "1.0.0"
description = "JSR Hydra Trading System"
requires-python = ">=3.11"
authors = [{name = "JSR Team"}]
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy[asyncio]",
    "asyncpg",
    "alembic",
    "pydantic-settings",
    "redis[hiredis]",
    "mt5linux",
    "structlog",
    "numpy",
    "pandas",
    "scikit-learn",
    "httpx",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-multipart",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "httpx",
    "fakeredis",
]
